AGE-67 migrate extract tests to tests/extract/ + extract public-API helpers to testing.py
Implements Option A1 from AGE-67: move all extract unit tests out of
src/groundx/extract/, separate public-API helpers from actual tests, and
reconcile the testpaths/CLI inconsistency.

## What changed

### New module: src/groundx/extract/classes/testing.py

Public test helpers previously embedded in test_field.py / test_groundx.py
move here. They remain re-exported as part of the groundx.extract public
API (no breaking change for downstream code that uses TestField,
TestChunk, TestDocumentPage, TestXRay).

Each helper class carries `__test__ = False` to tell pytest "this is not a
unittest class despite the Test* prefix" — eliminates the
PytestCollectionWarning the colocated layout generated.

### Test migration: src/groundx/extract/**/test_*.py → tests/extract/**/test_*.py

All 9 collocated test files move under tests/extract/, mirroring the
source structure. Relative imports become absolute
(from groundx.extract.X.Y import Z).

Cross-test fixtures (SAMPLE_YAML_1, SAMPLE_YAML_2, TestSource) extracted
to tests/extract/prompt/_fixtures.py — underscore prefix keeps pytest
from collecting the fixture module, TestSource carries __test__ = False.

Two internal test helpers renamed to avoid pytest collection confusion:
- test_xray() → _download_xray() in test_groundx.py
- test_doc() / test_request() → _make_doc() / _make_request() in test_document.py

### Infrastructure

- ci.yml: drop the positional `.` from pytest invocations. The existing
  testpaths = ["tests"] setting now applies cleanly; running
  `poetry run pytest` locally matches CI behavior.
- .fernignore: add tests/extract — protects the new test files from any
  future Fern template that touches tests/.

## Verified

- `poetry run mypy .` — Success, no issues found in 200 source files
- `poetry run pytest -rP -n auto` — 101 passed, 1 skipped, 1 pre-existing
  failure (TestExtractedField.test_get_value — production get_value()
  behavior diverged from the test expectation; identical failure on
  parent branch; not caused by this migration)
- Wheel-hygiene win: `pip install groundx[extract]` no longer ships
  ~75KB of test code to end users' site-packages
- Pytest collection warnings eliminated (TestSource, TestXRay,
  TestChunk, TestDocumentPage no longer cause PytestCollectionWarning)

## Public API

No breaking changes. `from groundx.extract import TestField, TestChunk,
TestDocumentPage, TestXRay` still works — the symbols are now sourced
from the new testing.py module via the existing
src/groundx/extract/classes/__init__.py re-export chain.
