+++ .fernignore
@@ -5,4 +5,5 @@ src/groundx/csv_splitter.py
 src/groundx/test.py
 src/groundx/extract
 src/test.py
+tests/extract
 .gitignore
